fix(qwik): ensure `data-frame-ui-spinner-leaf` attribute is rendered without a value
The change explicitly sets the data attribute `data-frame-ui-spinner-leaf` to an empty string (`data-frame-ui-spinner-leaf=""`).
This is necessary because the default rendering behavior in the absence of an assigned value was causing the attribute to be rendered as `data-frame-ui-spinner-leaf="true"` in the resulting HTML, which is undesirable for a presence attribute used for styling and selection.