fix(qwik): ensure `data-frame-ui-button-content` attribute is rendered without a value
The change explicitly sets the data attribute `data-frame-ui-button-content` to an empty string (`data-frame-ui-button-content=""`).
This is necessary because the default rendering behavior in the absence of an assigned value was causing the attribute to be rendered as `data-frame-ui-button-content="true"` in the resulting HTML, which is undesirable for a presence attribute used for styling and selection.